Docker/requirements.txt
azure-iot-device
asyncio
